pyproject.toml
[tool.poetry]
name = "langgraph develpment"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"
packages = [{ include = "langgraph develpment" }]  # Ensure a valid module exists

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "^0.3.18"
langchain-google-genai = "^2.0.9"
python-dotenv = "^1.0.1"
langgraph = "^0.2.70"
tavily-python = "^0.5.1"
langchain-community = "^0.3.17"
langchain-groq = "^0.2.4"
duckduckgo-search = "^7.3.2"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"